==> notebooks/01b_cohort_summary.ipynb <==
library(tidyverse)
library(RPostgreSQL)

sql_dir <- file.path("..", "SQL")
list.files(sql_dir)

drv <- dbDriver("PostgreSQL")
con <- dbConnect(drv, dbname = "mimic")
dbSendQuery(con, "set search_path=echo,public,mimiciii;")

#### Angus cohort

dbGetQuery(con, "select count(*) from population where angus = 1;")

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
"
dbGetQuery(con, sql)

#### Adult in their 1st ICU stay

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
   and icu_order = 1
"
dbGetQuery(con, sql)

#### MICU/SICU only

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
   and icu_order = 1
   and first_careunit in ('MICU', 'SICU')
"
dbGetQuery(con, sql)

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
   and icu_order = 1
   and first_careunit in ('MICU', 'SICU')
   and echo_exclude = 0
"
dbGetQuery(con, sql)

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
   and icu_order = 1
   and first_careunit in ('MICU', 'SICU')
   and echo_exclude = 1
"
dbGetQuery(con, sql)

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
   and icu_order = 1
   and first_careunit in ('MICU', 'SICU')
   and echo_include = 1
"
dbGetQuery(con, sql)

sql <- "
select count(*)
  from population
 where angus = 1
   and age >= 18
   and icu_order = 1
   and first_careunit in ('MICU', 'SICU')
   and echo_time is not null
   and echo_include = 1
"
dbGetQuery(con, sql)

dbDisconnect(con)|
dbUnloadDriver(drv)
