==> notebooks/02b_logi.ipynb <==
library(tidyverse)

data_dir <- file.path("..", "data")

model <- readRDS(file.path(data_dir, "multivariate_model.rds"))

coefs_mtx <- confint(model)

coefs <- coefs_mtx %>%
exp %>%
as.data.frame(stringsAsFactors = FALSE) %>%
setNames(c("lo", "hi")) %>%
mutate(cov = rownames(.))

ests <- model %>%
summary %>%
pluck("coefficients") %>%
as.data.frame(stringsAsFactors = FALSE) %>%
setNames(c("est", "se", "zval", "pval")) %>%
mutate(cov = rownames(.)) %>%
mutate(est = exp(est))

df <- coefs %>%
full_join(ests, by = "cov") %>%
mutate(ests = sprintf("%s (%s~%s)",
                      round(est, 2), round(lo, 2), round(hi, 2))) %>%
mutate(cov_ = cov) %>%
mutate(cov = case_when(
    grepl("icu_adm_hour|po2|pco2", cov) ~ cov,
    TRUE ~ str_replace_all(cov, "[0-9]", "")
)) %>%
mutate(cov = str_replace_all(cov, "icu_adm_weekday|icu_adm_hour", "")) %>%
mutate(cov = str_replace_all(cov, "_", " ")) %>%
mutate(cov = str_replace_all(cov, "lab|vs|flag|icd|first", "")) %>%
mutate(cov = str_replace_all(cov, "\\s+$|^\\s+", "")) %>%
mutate(cov = case_when(
    grepl("icd|lab|day", cov_) ~ toupper(cov),
    TRUE ~ tools::toTitleCase(cov)
)) %>%
mutate(cov = case_when(
    cov == "Echo" ~ "TTE",
    cov == "genderF" ~ "Gender (Female)",
    cov == "careunitMICU" ~ "Service Unit (MICU)",
    cov == "Weight" ~ "Weight (Kg)",
    cov == "Elix Score" ~ "Elixhauser Score",
    cov == "Saps" ~ "SAPS Score",
    cov == "Sofa" ~ "SOFA Score",
    cov == "Vent" ~ "Mechanical Ventilation (1st 24 Hr)",
    cov == "Vaso" ~ "Vasopressor (1st 24 Hr)",
    cov == "Sedative" ~ "Sedative (1st 24 Hr)",
    cov == "Icu Adm Weekday" ~ "Day of ICU Admission",
    cov == "Icu Adm Hour" ~ "Hour of ICU Admission",
    cov == "Temp" ~ "Temperature (C)",
    cov == "Map" ~ "MAP",
    cov == "Cvp" ~ "CVP (measured)",
    cov %in% c("TROPONIN", "BNP", "CREATININE KINASE") ~ paste(cov, "(tested)"),
    TRUE ~ cov
)) %>%
filter(!grepl("Intercept", cov)) %>%
mutate(pval = case_when(
    pval < 0.01 ~ "<0.01",
    TRUE ~ as.character(round(pval, 2))
)) %>%
select(cov, ests, pval) %>%
setNames(c("Covariate", "Odds Ratio (2.5%~97.5%)", "p value"))

df

data.table::fwrite(df, file.path(data_dir, "coefficients.csv"))
