==> notebooks/06_subgroup.ipynb <==
library(RPostgreSQL)
library(tidyverse)

source("utils.R")

sql_dir <- file.path("..", "sql")

drv <- dbDriver("PostgreSQL")
con <- dbConnect(drv, dbname = "mimic")
dbSendQuery(con, "set search_path=echo,public,mimiciii;")

files <- c("vitals", "labs", "blood-gas",
           "blood-gas-arterial", "gcs",
           "urine-output", "sofa") %>% paste0(".sql")
files

c("second_day", "third_day") %>%
walk(function(d) {
    files %>% walk(function(f) {
        cat(sprintf("-- Run %s from %s --\n", f, d))
        
        f %>%
        file.path(sql_dir, d, .) %>%
        readr::read_file() %>%
        dbSendQuery(con, .)
    })
})

cat("-- Generate fluid data --\n")

c("fluid_input", "fluid_output", "fluid") %>%
walk(function(f) {
    f %>%
    paste0(".sql") %>%
    file.path(sql_dir, .) %>%
    file_to_sql_view(f) %>%
    dbSendQuery(con, .)
})

cat("-- Generate lactate/creatinine reduction data --\n")

# c("serum_non_diff", "serum_echo_diff", "serum_diff") %>%
# walk(function(f) {
#     f %>%
#     paste0(".sql") %>%
#     file.path(sql_dir, .) %>%
#     file_to_sql_view(f) %>%
#     dbSendQuery(con, .)
# })

c("lactate_echo", "lactate_non_echo", "lactate_diff",
  "creatinine_echo", "creatinine_non_echo", "creatinine_diff") %>%
paste0(".sql") %>%
walk(function(f) {
    f %>%
    file.path(sql_dir, .) %>%
    readr::read_file() %>%
    dbSendQuery(con, .)
})

cat("-- Generate view subgroup --\n")

sql <- file_to_sql_view(file.path(sql_dir, "subgroup.sql"), "subgroup")
dbSendQuery(con, sql)

dbDisconnect(con)
dbUnloadDriver(drv)
