==> notebooks/07a_secondary_missing_data.ipynb <==
library(RPostgreSQL)
library(tableone)
library(tidyverse)

data_dir <- file.path("..", "data")

drv <- dbDriver("PostgreSQL")
con <- dbConnect(drv, dbname = "mimic")
dbSendQuery(con, "set search_path=echo,public,mimiciii;")

sql <-
"
select
sum(case when extract(day from (deathtime - intime)) < 1 then 1 else 0 end) as day1,
sum(case when extract(day from (deathtime - intime)) < 2 then 1 else 0 end) as day2
from merged_data
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.icustay_id, co.deathtime, co.intime, fl.day2, fl.day3
    from merged_data co left join fluid fl on co.icustay_id = fl.icustay_id
)

select
sum(case when extract(day from (deathtime - intime)) < 1 and day2 is not null then 1 else 0 end) as day2,
sum(case when extract(day from (deathtime - intime)) < 2 and day3 is not null then 1 else 0 end) as day3
from stg_0
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.icustay_id, co.deathtime, co.intime, fl.day1, fl.day2, fl.day3
    from merged_data co left join fluid fl on co.icustay_id = fl.icustay_id
)

select
sum(case when day1 is null then 1 else 0 end) as day1,
sum(case when day2 is null then 1 else 0 end) as day2,
sum(case when day3 is null then 1 else 0 end) as day3
from stg_0
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from merged_data where deathtime < intime + interval '3' day
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select * from merged_data left join fluid using (icustay_id)
)

select
sum(case when day1 is not null then 1 else 0 end) as day1,
sum(case when day2 is not null then 1 else 0 end) as day2,
sum(case when day3 is not null then 1 else 0 end) as day3
from stg_0 where deathtime < intime + interval '3' day
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from lactate_death
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.icustay_id, sr.lactate_diff
    from merged_data co left join lactate_diff sr on co.hadm_id = sr.hadm_id
)

select
sum(case when lactate_diff is null then 1 else 0 end) as lactate_diff
from stg_0
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.hadm_id, sr.lactate_diff
    from merged_data co left join lactate_diff sr on co.hadm_id = sr.hadm_id
)

select
sum(case when b.death = 1 and a.lactate_diff is not null then 1 else 0 end) as lactate_diff
from stg_0 a
left join (select hadm_id, 1 as death from lactate_death) b on a.hadm_id = b.hadm_id
"

dbGetQuery(con, sql)

sql <-
"
select count(*)
from lactate_diff inner join lactate_death using (hadm_id)
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from creatinine_death
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.icustay_id, sr.creatinine_diff
    from merged_data co left join creatinine_diff sr on co.hadm_id = sr.hadm_id
)

select
sum(case when creatinine_diff is null then 1 else 0 end) as creatinine_diff
from stg_0
"

dbGetQuery(con, sql)

sql <-
"
select count(*)
from creatinine_diff inner join creatinine_death using (hadm_id)
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.hadm_id, sr.creatinine_diff
    from merged_data co left join creatinine_diff sr on co.hadm_id = sr.hadm_id
)

select
sum(case when b.death = 1 and a.creatinine_diff is not null then 1 else 0 end) as creatinine_diff
from stg_0 a
left join (select hadm_id, 1 as death from creatinine_death) b on a.hadm_id = b.hadm_id
"

dbGetQuery(con, sql)

sql <- "select count(*) from lactate_discharge"

dbGetQuery(con, sql)

sql <-
"
select count(*) from lactate_diff inner join lactate_discharge using (hadm_id)
"

dbGetQuery(con, sql)

sql <- "select count(*) from creatinine_discharge"

dbGetQuery(con, sql)

sql <-
"
select count(*) from creatinine_diff inner join creatinine_discharge using (hadm_id)
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from lactate_death_24
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.icustay_id, sr.lactate_diff
    from merged_data co left join lactate_diff_24 sr on co.hadm_id = sr.hadm_id
)

select
sum(case when lactate_diff is null then 1 else 0 end) as lactate_diff
from stg_0
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from lactate_diff_24 inner join lactate_death_24 using (hadm_id)
"

dbGetQuery(con, sql)

sql <- "select count(*) from lactate_discharge_24"

dbGetQuery(con, sql)

sql <-
"
select count(*) from lactate_diff_24 inner join lactate_discharge_24 using (hadm_id)
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from creatinine_death_24
"

dbGetQuery(con, sql)

sql <-
"
with stg_0 as (
    select co.icustay_id, sr.creatinine_diff
    from merged_data co left join creatinine_diff_24 sr on co.hadm_id = sr.hadm_id
)

select
sum(case when creatinine_diff is null then 1 else 0 end) as creatinine_diff
from stg_0
"

dbGetQuery(con, sql)

sql <-
"
select count(*) from creatinine_diff_24 inner join creatinine_death_24 using (hadm_id)
"

dbGetQuery(con, sql)

sql <- "select count(*) from creatinine_discharge_24"

dbGetQuery(con, sql)

sql <-
"
select count(*) from creatinine_diff_24 inner join creatinine_discharge_24 using (hadm_id)
"

dbGetQuery(con, sql)

dbDisconnect(con)
dbUnloadDriver(drv)
