==> notebooks/10_subgroup_vaso_ps_matching.ipynb <==
library(Matching)
library(tidyverse)

data_dir <- file.path("..", "data")

full_data <- readRDS(file.path(data_dir, "full_data_ps_subgroup_vaso.rds"))
head(full_data)

set.seed(3485)

ps_matches <- Match(Y = full_data$mort_28_day_int, Tr = full_data$echo_int,
                    X = full_data$ps, M = 1, estimand = "ATT", caliper = 0.01,
                    exact = FALSE, replace = FALSE, ties = FALSE)

ps_matches_df <-
full_data[ps_matches$index.treated, "icustay_id", drop = FALSE] %>%
mutate(match = full_data[ps_matches$index.control, "icustay_id"]) %>%
rename(trtd = icustay_id, ctrl = match) %>%
mutate(match = 1:n()) %>%
gather("group", "icustay_id", trtd, ctrl) %>%
select(icustay_id, group, match) %>%
arrange(group, match)

head(ps_matches_df)

# saveRDS(ps_matches, file = file.path(data_dir, "ps_matches.rds"))
data.table::fwrite(ps_matches_df, file.path(data_dir, "ps_matches_df_subgroup_vaso.csv"))

ids <- full_data[unlist(ps_matches[c("index.control", "index.treated")]), "icustay_id"]
length(ids) / 2
head(ids)

data.table::fwrite(data.frame(icustay_id = ids), file = file.path(data_dir, "ps_icustay_id"))

summary(ps_matches)

outcome <- data.frame(echo_pt = full_data$mort_28_day_int[ps_matches$index.treated],
                      match_pt = full_data$mort_28_day_int[ps_matches$index.control])

head(outcome)

tab <- table(outcome$echo_pt, outcome$match_pt, dnn = c("Echo", "Non-Echo"))
tab
tab[2, 1] / tab[1, 2]
paste("95% Confint",
      round(exp(c(log(tab[2, 1] / tab[1, 2]) - qnorm(0.975) * sqrt(1 / tab[1, 2] + 1 / tab[2, 1]),
                  log(tab[2, 1] / tab[1, 2]) + qnorm(0.975) * sqrt(1 / tab[1, 2] + 1 / tab[2, 1]))), 7))
mcnemar.test(tab) # for 1-1 pairs

(tab[2, 2] + tab[1, 2]) / sum(tab)
(tab[2, 2] + tab[2, 1]) / sum(tab)

mortality <- data.frame(echo = c(0, 1),
                        mortality = c((tab[2, 2] + tab[1, 2]) / sum(tab),
                                      (tab[2, 2] + tab[2, 1]) / sum(tab)))
mortality

mort_fmt <- mortality %>%
mutate(echo = case_when(
    echo == 0 ~ "Non_echo",
    echo == 1 ~ "Echo"
)) %>%
select(echo, mortality) %>%
spread(echo, mortality) %>%
mutate(Method = "PS Match") %>%
select(Method, Echo, Non_echo)
mort_fmt

data.table::fwrite(mort_fmt, file.path(data_dir, "mort_ps_subgroup_vaso.csv"))
