==> notebooks/13_subgroup_vaso_mortality.ipynb <==
library(RPostgreSQL)
library(tidyverse)

data_dir <- file.path("..", "data")
sql_dir <- file.path("..", "sql")

drv <- dbDriver("PostgreSQL")
con <- dbConnect(drv, dbname = "mimic")
dbSendQuery(con, "set search_path=echo,public,mimiciii;")

sql <-
"
with mort_0 as (
    select echo,
        sum(case when mort_28_day = 1 then 1 else null end) as death,
        count(*) as total
    from merged_data
    where vaso = 1
    group by echo
)

, mort as (
    select echo, death, total, death / 1.0 / total as mortality
    from mort_0
)

select * from mort
"

mortality <- dbGetQuery(con, sql)
mortality

mort_fmt <- mortality %>%
mutate(echo = case_when(
    echo == 0 ~ "Non_echo",
    echo == 1 ~ "Echo"
)) %>%
select(echo, mortality) %>%
spread(echo, mortality) %>%
mutate(Method = "Raw") %>%
select(Method, Echo, Non_echo)
mort_fmt

data.table::fwrite(mort_fmt, file.path(data_dir, "mort_raw_subgroup_vaso.csv"))

mort_ps <- data.table::fread(file.path(data_dir, "mort_ps_subgroup_vaso.csv"))

mort_cmb <- rbind(mort_fmt, mort_ps) %>% setNames(c("Method", "Echo", "Non-Echo"))
mort_cmb

data.table::fwrite(mort_cmb, file.path(data_dir, "mortality_subgroup_vaso.csv"))

dbDisconnect(con)
dbUnloadDriver(drv)
