==> notebooks/14_subgroup_vaso_infection.ipynb <==
library(RPostgreSQL)
library(tidyverse)

data_dir <- file.path("..", "data")
sql_dir <- file.path("..", "sql")

drv <- dbDriver("PostgreSQL")
con <- dbConnect(drv, dbname = "mimic")
dbSendQuery(con, "set search_path=echo,public,mimiciii;")

sql <-
"
with infection_0 as (
    select hadm_id,
        max(case when icd9_code in ('480', '481', '482', '483', '484', '485', '486', '487', '488') then 1 else 0 end) as infection
    from diagnoses_icd
    group by hadm_id
)

, infection_1 as (
    select *
    from merged_data
    left join infection_0 using (hadm_id)
)

, infection as (
    select icustay_id, echo as echo, coalesce(infection, 0) as infection
    from infection_1
)

select * from infection;
"

infection <- dbGetQuery(con, sql)
str(infection)
head(infection)

infection %>%
group_by(echo) %>%
summarise(n = n(), infection = sum(infection)) %>%
mutate(p = infection / n)

chisq.test(infection$infection, infection$echo)

ps_matches_df <- data.table::fread(file.path(data_dir, "ps_matches_df.csv"), data.table = FALSE)

head(ps_matches_df)

ps_df <-
ps_matches_df %>%
left_join(infection, by = "icustay_id")

head(ps_df)

matched <-
ps_df %>%
filter(echo == 1) %>%
rename(trtd = infection) %>%
inner_join(ps_df %>%
           filter(echo == 0) %>%
           rename(ctrl = infection), by = "match")

head(matched)

tab <- table(matched$trtd, matched$ctrl, dnn = c("Echo", "Non-Echo"))
tab

tab[2, 1] / tab[1, 2]
round(exp(c(log(tab[2, 1] / tab[1, 2]) - qnorm(0.975) * sqrt(1 / tab[1, 2] + 1 / tab[2, 1]),
            log(tab[2, 1] / tab[1, 2]) + qnorm(0.975) * sqrt(1 / tab[1, 2] + 1 / tab[2, 1]))), 7)

mcnemar.test(tab)

dbDisconnect(con)
dbUnloadDriver(drv)
